Code formatting: Semantic linebreaks and indent
This change doesn't affect the LaTeX output.  The only change is to
switch to semantic linebreaks for the code [1] and use consistent LaTeX code
style [2].  The reasoning behind that can be read under the linked
pages.  Specifically note that I try to put URLs on their separate lines
to make reading and changes to the code easy.

[1]: https://gitlab.it-sec.intern/pentest/reportlatex/itsec-latex/-/issues/254
[2]: https://mirror.mwt.me/ctan/macros/latex/contrib/l3kernel/l3styleguide.pdf