Use LaTeX \(...\) instead of TeX $...$
The IMHO biggest upside is better error messages as, e.g., the following

	$3+4 baz meep $5+3$ meow
         ^^^^^^^^^^^^^     ^^^^^^^

ambiguous to LaTeX:  Where is the missing $?  Is it after the meow and
extending to the end of document (hint: no), or is it actually ending
after the "3+4" (hint: yes).  With the `\(...\)`-Notation we introduce a
difference b/w opening and closing math mode, so error messages can
become much more helpful!

	\(3+4 baz meep \(5+3\) meow
	  ^^^^^^^^^^^^^  ^^^

Now it can recognize 5+3 as an intact math environment and conclude that
the missing closing math mode is in the text before!